requirements.txt
fastapi
uvicorn
requests
python-decouple
